LPS-104271 Refactor how props are managed in some components
Now properties that come from Store are computed inside
mapStateToProps callback, which produces a simple state easier
to read by metal.

This produces a much more efficient behavior when changing properties
like hoveredItemId and activeItemId, because this components
are not reaching the render lifecycle method.

Although this might be refactored in all ConnectedComponents, we are
only refactoring FragmentEntryLink and EditableField, as those are
the two components suffering many unnecessary re-renders. We may
refactor more components to use this new method as performance issues
appear, or until we can finish our React migration